+++ project_task_subtask/doc/changelog.rst
@@ -3,6 +3,7 @@
 -------
 
 - **Improvement**: 'Checklist' replaced by 'To-Do List'
+- **New**: added group "Allow edit others tasks"
 
 `1.0.0`
 -------